+++ NEWS.md
@@ -7,6 +7,7 @@
 ## Other improvements
 
 * Added a `NEWS.md` file to track changes to the package.
+* `check_downstream` will not stop on first package error but will run on all expected packages and all failures are output. 
 
 ## Bugfixes